Change how we generate tokens
`passport-http-bearer` doesn't have its own token generation method, so
we had to improvise. Instead, we just generate 16 random bytes as a
token. As noted in the comments, this is not production-ready security,
but it's good enough for pedagogical purposes, and has the advantage of
being really simple and understandable.

Also caught a bug where incorrect passwords were accepted, because
pending promises are truthy. Whoops.